pyproject.toml
[project]
name = "chat-with-archive"
version = "0.1.0"
description = "Chat with an AI question-answering agent that has tools for getting Twitter Community Archive data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "jinja2>=3.1.5",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.12.0",
    "types-requests>=2.32.0.20241016",
]
